+++ spec/services/reports/csv_reports_service_spec.rb
@@ -5,11 +5,7 @@ RSpec.describe Reports::CsvReportsService do
     described_class.new
   end
 
-  let(:form_documents_url) { "#{Settings.forms_api.base_url}/api/v2/form-documents".freeze }
-  # This response JSON was generated by making a real API request to forms-api with the data from the database seeds.
-  # Once we have transitioned to using the V2 API in forms-admin, it might make more sense to use factories to generate
-  # the response.
-  let(:form_documents_response_json) { file_fixture("form_documents_response.json").read }
+  let(:form_documents_response_json) { JSON.parse(file_fixture("form_documents_response.json").read) }
 
   let(:group) { create(:group) }
 
@@ -18,29 +14,14 @@ RSpec.describe Reports::CsvReportsService do
     GroupForm.create!(form_id: 2, group:)
     GroupForm.create!(form_id: 3, group:)
 
-    stub_request(:get, form_documents_url)
-      .with(query: { page: "1", per_page: "3", tag: "live" })
-      .to_return(body: file_fixture("form_documents_response.json").read, headers: response_headers(9, 0, 3))
-    stub_request(:get, form_documents_url)
-      .with(query: { page: "2", per_page: "3", tag: "live" })
-      .to_return(body: file_fixture("form_documents_response.json").read, headers: response_headers(9, 3, 3))
-    stub_request(:get, form_documents_url)
-      .with(query: { page: "3", per_page: "3", tag: "live" })
-      .to_return(body: file_fixture("form_documents_response.json").read, headers: response_headers(9, 6, 3))
+    allow(Reports::FormDocumentsService).to receive(:live_form_documents).and_return(form_documents_response_json)
   end
 
   describe "#live_forms_csv" do
-    it "makes request to forms-api for each page of results" do
-      csv_reports_service.live_forms_csv
-      assert_requested(:get, form_documents_url, query: { page: "1", per_page: "3", tag: "live" }, headers:, times: 1)
-      assert_requested(:get, form_documents_url, query: { page: "2", per_page: "3", tag: "live" }, headers:, times: 1)
-      assert_requested(:get, form_documents_url, query: { page: "3", per_page: "3", tag: "live" }, headers:, times: 1)
-    end
-
-    it "returns a CSV with 10 rows, including the header row" do
+    it "returns a CSV with 4 rows, including the header row" do
       csv = csv_reports_service.live_forms_csv
       rows = CSV.parse(csv)
-      expect(rows.length).to eq 10
+      expect(rows.length).to eq 4
     end
 
     it "has expected values" do
@@ -69,34 +50,13 @@ RSpec.describe Reports::CsvReportsService do
         "email",
       ])
     end
-
-    context "when forms-api responds with a non-success status code" do
-      before do
-        stub_request(:get, form_documents_url)
-          .with(query: { page: "1", per_page: "3", tag: "live" })
-          .to_return(body: "There was an error", status: 400)
-      end
-
-      it "raises a StandardError" do
-        expect { csv_reports_service.live_forms_csv }.to raise_error(
-          StandardError, "Forms API responded with a non-success HTTP code when retrieving form documents: status 400"
-        )
-      end
-    end
   end
 
   describe "#live_forms_questions" do
-    it "makes request to forms-api for each page of results" do
-      csv_reports_service.live_questions_csv
-      assert_requested(:get, form_documents_url, query: { page: "1", per_page: "3", tag: "live" }, headers:, times: 1)
-      assert_requested(:get, form_documents_url, query: { page: "2", per_page: "3", tag: "live" }, headers:, times: 1)
-      assert_requested(:get, form_documents_url, query: { page: "3", per_page: "3", tag: "live" }, headers:, times: 1)
-    end
-
-    it "returns a CSV with 46 rows, including the header row" do
+    it "returns a CSV with 16 rows, including the header row" do
       csv = csv_reports_service.live_questions_csv
       rows = CSV.parse(csv)
-      expect(rows.length).to eq 46
+      expect(rows.length).to eq 16
     end
 
     it "has expected values for text question" do
@@ -214,27 +174,5 @@ RSpec.describe Reports::CsvReportsService do
         "{\"only_one_option\" => \"true\", \"selection_options\" => [{\"name\" => \"Once\"}, {\"name\" => \"More than once\"}]}",
       ])
     end
-
-    context "when forms-api responds with a non-success status code" do
-      before do
-        stub_request(:get, form_documents_url)
-          .with(query: { page: "1", per_page: "3", tag: "live" })
-          .to_return(body: "There was an error", status: 400)
-      end
-
-      it "raises a StandardError" do
-        expect { csv_reports_service.live_questions_csv }.to raise_error(
-          StandardError, "Forms API responded with a non-success HTTP code when retrieving form documents: status 400"
-        )
-      end
-    end
-  end
-
-  def response_headers(total, offset, limit)
-    {
-      "pagination-total" => total.to_s,
-      "pagination-offset" => offset.to_s,
-      "pagination-limit" => limit.to_s,
-    }
   end
 end
